refactor : Week6 SOLID
